Convert to TS many tests
