+++ src/prec-0/Prec0DatasetUtil.ts
@@ -3,7 +3,7 @@
  * Methods related to consuming an RDF graph generated by PREC-0
  */
 
-import { DatasetCore, Quad_Object, Quad_Subject, Term } from "@rdfjs/types";
+import { DatasetCore, Quad_Object, Quad_Subject } from "@rdfjs/types";
 
 import { DataFactory } from "n3";
 import namespace from '@rdfjs/namespace';
