+++ test/prec_impl/test-function.d.ts
@@ -1,0 +1,2 @@
+
+export type PrecCApplicationTester = (testName: string, prec0Graph: string, context: string, expected: string) => void;
