--- test/utility.js
@@ -1,29 +1,0 @@
-const N3 = require("n3");
-const { default: DStar } = require('../src/dataset/index');
-
-function readQuads(turtleContent) {
-    const prefixes =
-    `
-        @prefix     : <http://test/>                                .
-        @prefix rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
-        @prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#>       .
-        @prefix prec: <http://bruy.at/prec#>                        .
-        @prefix pgo:  <http://ii.uwb.edu.pl/pgo#>                   .
-        @prefix pvar: <http://bruy.at/prec-trans#>                  .
-    `;
-
-    const parser = new N3.Parser();
-
-    try {
-        let p = parser.parse(prefixes + turtleContent);
-        return p;
-    } catch (e) {
-        console.error("Passed turtle file is not valid\n" + turtleContent);
-        throw e;
-    }
-}
-
-module.exports = {
-    turtleToQuads: readQuads,
-    turtleToDStar: content => new DStar(readQuads(content)),
-};
